### 2023-07-19 #5666 @editorialbot
[PRE REVIEW]: prospect: an R package to link leaf optical properties with their chemical and structural properties with the leaf model PROSPECT
**Submitting author:** @jbferet (<a href="http://orcid.org/0000-0002-0151-1334">Jean-Baptiste Feret</a>)
**Repository:** https://github.com/jbferet/prospect
**Branch with paper.md** (empty if default branch):
**Version:** v1.3.0
