### 2023-08-16 e62fb65
v1.4.0 manages value of LMA when set to NULL or 0, depending on when PROSPECT- D or PROSPECT-PRO is expected
