### 2023-11-07 #6027 @marcosvital
Dear @earth-chris, @nehakn and @ashiklom, thank you again for accepting review this submission for JOSS.
The reviewing process is checklist based, and instructions were already posted above by the editorial bot - but let me know if you need any assistance, ok? Also, you can tag  @jbferet if you have specific questions about the manuscript.

@jbferet, you can tag your co-author GitHub accounts if you want, so they will be able to follow this issue and answer to questions as well.
