### 2023-11-21 jbferet/prospect#3 @ashiklom
Consider a different license
(Commenting as part of review for https://github.com/openjournals/joss-reviews/issues/6027)

I understand the rationale for the GPL license, and I recognized that it's approved by the Open Source Initiative (OSI). However, because of its strong copy-left provisions, the GPL license will make this library difficult to integrate into other software because that software _also_ has to be licensed under GPL. This means that the library can never (legally) be integrated into code with any other license (though I think it can be _called_ from non-GPL-licensed code...? See, for example, recent discussions around this in the Julia community: https://github.com/JuliaLang/julia/issues/45833).

In particular, I would love to see this implementation of PROSPECT used by, e.g., NASA missions and field campaigns. However, the GPL license may make developers jump through extra hoops to do that, _and_ will cause NASA lawyers a lot of headache. I strongly suspect that some projects will find it easier to roll their own PROSPECT implementation than try to deal with the issues of integrating a GPL-licensed library into their workflow.

Good alternative licenses without these issues are Apache 2.0 or MIT.

Just something to consider! If you have a compelling reason to keep this as GPL v3, that's absolutely your right, but just a heads up that doing so may reduce usage of this library for some use cases.
