### 2023-11-21 jbferet/prospect#4 @ashiklom
Consider including plain-text versions of PROSPECT coefficients
(Comment as part of JOSS review: https://github.com/openjournals/joss-reviews/issues/6027)

I think the typical best practice for R packages is to include plain-text versions of internal package data and the code used to convert them into the internal data. Here's the [relevant discussion in Writing R Packages](https://r-pkgs.org/data.html#sec-data-sysdata). In the context of this package, you might consider also including plain-text versions of the PROSPECT absorption coefficients and refractive index and the code used to put these into the final PROSPECT input data frame (which is stored in binary representation in the package). Among other things, this will (1) allow you to version-control the PROSPECT coefficients in human-readable git fashion (e.g., in case the calibrations change in the future), and (2) make the package resilient to changes in the binary encoding of `Rdata` and `rda` files.

This is **not** a requirement for my review; just a friendly suggestion!
