### 2023-11-21 jbferet/prospect#5 @ashiklom
Move randomcoloR to Suggests
`randomcoloR` depends on `V8`, which requires a system library (V8 engine). That's a significant dependency just to get more pleasing colors in one visualization in the package.

I suggest the following:

(1) Move `randomcoloR` to `Suggests`.

(2) Remove the `importFrom randomcoloR distinctColorPalette` from `R/Lib_PROSPECT_Inversion.R`.

(3) Replace the relevant part of the `plotinv` code with something like this:

```r
  if (is.null(Colors)) {
    if (requireNamespace("randomcoloR")) {
      # If `randomcoloR` is installed, use it.
      Colors <- randomcoloR::distinctColorPalette(unique(BP_df$config))
    } else {
    # If not, fall back on the built-in (but less pleasing) rainbow color palette (or some other default palette)
      Colors <- rainbow(unique(BP_df$config))
    }
  }
```

This will make the package much easier to install!

---

More generally, it's a good idea to only _require_ dependencies that are essential to the core functioning of the package (in this case, things like `expint` and the libraries used for the inversion) and _suggest_ stuff that is useful for _demonstrating_ capabilities of the package (e.g., `ggplot` for inversion plotting; `data.table` for loading example data).
