### 2023-11-21 jbferet/prospect#6 @ashiklom
Suggestion: Pre-calculate `calctav`
These values in PROSPECT are constants that do not depend on the PROSPECT parameters:

https://github.com/jbferet/prospect/blob/5c43c1b2a72a2f0fb92cd47011ea5da31add58e5/R/Lib_PROSPECT.R#L88-L93

You can squeeze some extra performance out of PROSPECT by storing these in your `Input_PROSPECT` data frame. This won't be noticeable for single PROSPECT simulations, but might matter a bit more for inversion and LUT generation.

If you follow the recommendation in #4 to generate `Input_PROSPECT` in a script that lives inside this repository, you can move this code into there so it is reproducible (and so you don't forget how these quantities are generated😅).

(Part of https://github.com/openjournals/joss-reviews/issues/6027).
