### 2023-11-21 jbferet/prospect#7 @ashiklom
Add automated unit tests
I suggest adding automated unit tests of core functionality (PROSPECT simulations and inversion). For R packages, the best practice is to use testthat, [as described in Writing R Packages](https://r-pkgs.org/testing-basics.html).

For PROSPECT simulations, something simple like:

```r
test_that("PROSPECT-D produces physically possible values", {
  lrt <- PROSPECT(SpecProspect, ...)
  expect_true(all(lrt$Reflectance >= 0))
  expect_true(all(lrt$Reflectance <= 1))
  expect_true(all(lrt$Transmittance >= 0))
  expect_true(all(lrt$Transmittance <= 1))
}
```

For inversion, I typically suggest (1) running PROSPECT in forward mode to generate reflectance for a known parameter value; (2) add a bit of random noise; (3) run an inversion on the simulated data; (4) check that the inversion estimate is within some error margin of the known parameter value.

You automate testing via GitHub Actions workflows using [r-lib/actions](https://github.com/r-lib/actions).

(Part of JOSS review: https://github.com/openjournals/joss-reviews/issues/6027).
