### 2023-11-21 jbferet/prospect#8 @ashiklom
Add contribution guide
Either add a dedicated `CONTRIBUTING.md` file (e.g., https://contributing.md/how-to-build-contributing-md/), or --- probably easier --- add a brief "Contributing" section to the `README.md` that includes the following (pulled directly from JOSS guidelines):

> "clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support"

(Part of JOSS review https://github.com/openjournals/joss-reviews/issues/6027)
