### 2023-11-21 jbferet/prospect#9 @ashiklom
Paper: Mention implementation history (briefly)
This is a JOSS requirement:

> **State of the field:** Do the authors describe how this software compares to other commonly-used packages?

In the paper, you talk about the history of PROSPECT development. I would also just add 1-2 more sentences on existing implementations of PROSPECT; basically, everything on your [teledetection website](http://teledetection.ipgp.jussieu.fr/prosail/), plus maybe this Python implementation (https://github.com/jgomezdans/prosail) and this Julia implementation (https://github.com/RemoteSensingTools/CanopyOptics.jl).

(Just for your reference, there's [also my R implementation](https://github.com/ashiklom/rrtm), but I'm not sure if anyone other than me and 1-2 of my colleagues actually uses it and I haven't touched it in a while. Also, I consider it bad form when reviewers request their own papers be cited, so no need to cite it explicitly --- just know that it exists and feel free to use it for inspiration 😄).

Related to JOSS review https://github.com/openjournals/joss-reviews/issues/6027.
