### 2023-12-13 #6027 @kthyng
@ashiklom and @jbferet A note — automated testing is not required by JOSS, but *some means* of testing that the code is running correctly is.

With that in mind, @jbferet how are your edits coming along?

@earth-chris, @nehakn How are your reviews coming?
