### 2023-12-17 jbferet/prospect#5 @earth-chris
I'd like to second this recommendation. I spent a good chunk of time figuring out how to install V8 (and found [this set of solutions](https://stackoverflow.com/questions/62302713/not-able-to-install-v8-package-on-ubuntu)). If you plan to keep this is a required dependency, I might recommend providing links or tips on how to install this dependency.
