### 2023-12-17 jbferet/prospect#3 @earth-chris
I'd like to second this recommendation. GPL licenses limit the software many organizations are willing to use, and a restrictive license could lead users to prefer other prospect implementations over your "canonical" version
