### 2023-12-17 jbferet/prospect#11 @earth-chris
REVIEW: manuscript feedback
hi @jbferet! here's my manuscript feedback.

High level:

- My understanding is that this manuscript is designed to point the research community to the primary implementation of the most modern PROSPECT code.
- I recommend you clarify the long-term intent of this R package (because the different, multi-language versions can be quite confusing for users). Will new versions be implemented in this package, or in another repository? Will future updates still be in R? Will old versions or versions in other languages be deprecated?
- The inclusion of versions -DB and -PROB in Table 1 is also confusing. The manuscript states that this package includes both versions -D and -PRO (lines 38-39), but the functions support the `BROWN =` argument, indicating that it actually includes -DB and -PROB. I recommend either updating the version you report is used by this package, or dropping the B versions altogether.
- As a user, I'd like to have a better understanding of what the limitations of the software are. Can this be easily used to simulate or invert spectra from multispectral satellite data? is the default wavelength specification (1nm band widths, full visible-to-swir) required? what is the scale at which inversions can be applied (hundreds, thousands, millions of spectra)?
- I recommend prioritizing figures and tables over code blocks in the manuscript. The code blocks do not tell us much about what the code does or how it does it. These blocks also include a lot of comments, array creation, or file downloads, which are not relevant to the manuscript. These are better off in the web docs.

Line comments:

- lines 24-26: "[the package] includes parameterizable inversion routines, allowing users to design and test their own inversion strategy." how can users design and test their own inversion strategy? this is not documented in the manuscript or the web docs.
- lines 60-65: a LUT simulation is provided but not used elsewhere in the manuscript or the web docs. It's not clear what this does, or why it is useful. An example figure, code snippet, or citation showing how LUTs are used would help.
- line 73: "various inversion strategies can be found in the literature." please add citations for this.
- line 82: the ANGERS download code block is extra bulky and not particularly relevant to the manuscript. i'd recommend either creating simple helper function to download this data or removing the code blocks from this section entirely.
- lines 106-111: the conclusions state that PROSPECT can be run with bidirectional or reflectance/transmission-only data. This seems like an important point to make, and you might clarify the pros and cons of each approach. Reading the manuscript and the examples, I assumed you needed bidirectional data.

Overall, I'm glad to see this new version of PROSPECT, and excited as a user to use the inversion methods. I think you have an opportunity to provide some needed clarity around the existing PROSPECT software ecosystem here, which I don't believe has been addressed in prior manuscripts. Cheers.

(Part of JOSS review https://github.com/openjournals/joss-reviews/issues/6027)
