### 2023-12-17 #6027 @earth-chris
hi @jbferet, I've concluded my review. Here's my recommended [manuscript input](https://github.com/jbferet/prospect/issues/11). I have left a couple of boxes unchecked but I think there is a short pathway to resolving my current issues.

First is that the install instructions are incomplete. There are external dependencies that need to be installed (V8) that are not documented, which I had to debug. I've chimed in with an additional comment on [ashiklom's issue](https://github.com/jbferet/prospect/issues/5).

Second is that the examples provided in the web documentation don't all run as-is. Primarily, the very first code block doesn't run. There's also an opportunity to port the more recent code
blocks from the manuscript to the web docs, which would provide more comprehensive documentation. I've described that in more detail in [this issue](https://github.com/jbferet/prospect/issues/10).

I would like to +1 ashiklom's review, and believe this software will be much improved after addressing those key issues. Many of them are nice-to-haves (like automated testing), and the only blockers to me are the install issues and the broken code in the web docs.
