### 2023-12-18 jbferet/prospect#3 @jbferet
thank you @ashiklom and @earth-chris for the suggestion.
I will change the license in the release in preparation following your reviews. I will most probably use an MIT license.
