### 2023-12-18 jbferet/prospect#5 @jbferet
sorry for the waste of time @earth-chris ! and thank you @ashiklom for noticing this issue. This function is actually not strictly needed, and I could not find any direct use in the current version of the package.
so I will remove it from the next release. It will make it easier, and reduce dependency to other packages.
