### 2023-12-20 jbferet/prospect#6 @jbferet
Excellent suggestion, thank you @ashiklom !
This also substancially decreased the processing time for inversion. I will include the outputs of calctav for solid angle of 90 degrees and 40 degrees (default) in the SpecPROSPECT dataframe, so its spectral range and sampling will be adjusted along with specific absorption coefficient when needed.
