### 2024-01-02 4c7f52d
# prospect v1.5.0 version following reviews for publication in JOSS
## Changes
- change license to MIT (Issue#3:  Consider a different license)
- include plain text version of specific absorption coefficients with variable prospect::SpecPROSPECT_FullRange (Issue#4:  Consider including plain-text versions of PROSPECT coefficients)
- remove function plotinv and corresponding required packages (Issue#5: Move randomcoloR to Suggests)
- pre-calculate calctav for angles 40 degrees (default) and 90 degrees (always needed) and add them to SpecPROSPECT_FullRange (Issue#6: Pre-calculate calctav)
- added
[…]
