+++ paper/paper.md
@@ -24,157 +24,221 @@ bibliography: paper.bib
 
 # Summary
 
-PROSPECT simulates leaf optical properties (LOP)
-based on a limited set of light absorbing chemical constituents, and a unique leaf 
-structure parameter to account for light scattering.
-We present `prospect`, an R package which includes the most recent versions of the model, 
-complemented with multiple inversion routines to assess leaf chemistry from LOP.
+PROSPECT simulates leaf optical properties based on a limited set of light 
+absorbing chemical constituents, and a unique leaf structure parameter to 
+account for light scattering.
+We present `prospect`, an R package which includes the most recent versions 
+of the model, complemented with multiple inversion routines to assess leaf 
+chemistry from leaf optical properties.
 
 # Statement of need
 
-LOP are linked to their biophysical properties through absorption 
-and scattering mechanisms. PROSPECT takes advantage of specific absorption coefficients from chemical constituents 
-such as leaf pigments, water and proteins, and uses a simplified representation of leaf structure 
-through the generalised plate model [@allen1970] to simulate leaf directional-hemispherical 
-reflectance and transmittance. 
-The capacity to measure, map and monitor vegetation chemical traits from leaf to canopy scale 
-is key to better understand ecosystems and agrosystems functions, as well as carbon, water and energy budgets. 
-Models and methods taking advantage of optical measurements to assess these core constituents are 
-important components in this perspective.
+The capacity to measure, map and monitor vegetation traits corresponding to 
+biophysical and chemical properties is crucial to better understand ecosystem 
+and agrosystem functions, as well as carbon, water and energy budgets. 
+At leaf scale, these vegetation traits are linked to their 
+optical properties through absorption and scattering mechanisms. 
+Techniques based on spectroscopy have developed in the past decades to provide 
+rapid, accurate and non-destructive assessment of leaf chemical composition. 
+Physical models aim at simulating optical properties of leaves from their 
+chemical composition, and they are also used to assess chemical composition 
+from leaf optical properties based on inversion techniques. 
+The model PROSPECT (leaf optical PROperties SPECtra) is currently the most popular 
+physical model used for the simulation of leaf optical properties. 
+PROSPECT simulates leaf directional-hemispherical reflectance and transmittance 
+from a combination of chemical constituents and their corresponding specific 
+absorption coefficients. 
+It uses a simplified representation of leaf structure through the generalised 
+plate model [@allen1970] to account for scattering. 
+PROSPECT is also coupled with canopy reflectance models to analyze Earth 
+observation data, such as satellite imagery. Hence, it is a key component for 
+remote sensing applications dedicated to vegetation monitoring. 
 
-Techniques based on spectroscopy have developed in the past decades to provide a 
-rapid, accurate and non-destructive estimation of leaf chemical composition. 
-Physical models aim at simulating LOP corresponding to this chemical composition, but they can 
-also be used to assess chemical composition from LOP based on inversion techniques. 
+Multiple versions aiming at expanding the pool of chemical constituents accounted 
+for by PROSPECT have been released since its first version [@jacquemoud1990].
[…]
