+++ R/OptDomain_RT-data.R
@@ -1,0 +1,17 @@
+#' OptDomain_RT: optimal spectral domains defined to assess the different leaf
+#' chemical constituents.
+#' These optimal spectral domains were defined in
+#' Féret et al. (2019) https://doi.org/10.1016/j.rse.2018.11.002
+#' Féret et al. (2021) https://doi.org/10.1016/j.rse.2020.112173
+#' Spafford et al. (2021) https://doi.org/10.1016/j.rse.2020.112176
+#'
+#' @format A list including sequence of spectral bands corresponding to each
+#' leaf constituent
+#'  - *CHL*: optimal spectral domain for CHL
+#'  - *CAR*: optimal spectral domain for CAR
+#'  - *ANT*: optimal spectral domain for ANT
+#'  - *EWT*: optimal spectral domain for EWT
+#'  - *LMA*: optimal spectral domain for LMA
+#'  - *PROT*: optimal spectral domain for PROT
+#'  - *CBC*: optimal spectral domain for CBC
+"OptDomain_RT"
