### 2024-01-03 545a7ca
minor update - update CI - update news - update paper
