+++ .github/workflows/R-CMD-check.yml
@@ -2,9 +2,9 @@
 # Need help debugging build failures? Start at https://github.com/r-lib/actions#where-to-find-help
 on:
   push:
-    branches: [main]
+    branches: [main, master]
   pull_request:
-    branches: [main]
+    branches: [main, master]
 
 name: test-coverage
 
@@ -13,28 +13,24 @@ jobs:
     runs-on: ubuntu-latest
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
-    name: R
+
     steps:
       - uses: actions/checkout@v3
-      - name: Setup R
-        uses: r-lib/actions/setup-r@v2
+
+      - uses: r-lib/actions/setup-r@v2
         with:
           use-public-rspm: true
+
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
-          extra-packages: |
-            any::covr
-            any::rcmdcheck
-          needs: |
-            coverage
-            check
-          working-directory: 'prospect'
+          extra-packages: any::covr
+          needs: coverage
+
       - name: Test coverage
         run: |
           covr::codecov(
             quiet = FALSE,
             clean = FALSE,
-            path = "prospect",
             install_path = file.path(Sys.getenv("RUNNER_TEMP"), "package")
           )
         shell: Rscript {0}
