### 2024-01-03 #6027 @jbferet
@kthyng yes I am!
Sorry for the poor communication, I have been working on corrections and updates of my package, following reviews from @ashiklom , @earth-chris and @nehakn .
I just pushed an updated version of the package and paper.
I will now exchange with @ashiklom and @earth-chris for specific suggestions they made in their detailed review listed in the issues of the package.
thanks!
