### 2024-01-04 jbferet/prospect#3 @jbferet
hi @ashiklom and @earth-chris !
I updated the package yesterday.
prospect v1.5.0 is now under MIT license.
