### 2024-01-04 jbferet/prospect#4 @jbferet
prospect v1.5.0 now includes additional information in the 'data-raw' directory:

1. [dataSpec_PRO.txt](https://github.com/jbferet/prospect/blob/master/data-raw/dataSpec_PRO.txt) includes the refractive index, specific absorption coefficients for each constituent, and corresponding wavelength with 1 nm sampling.

2. [create_sysdata.R](https://github.com/jbferet/prospect/blob/master/data-raw/create_sysdata.R) is the R script generating the variable _SpecPROSPECT_FullRange_, and the variable _OptDomain_RT_, available when loading the package.

- SpecPROSPECT_FullRange includes refractive index and specific absorption coefficient from dataSpec_PRO.txt, as well as the transmissivity computed with the function calctav for two angles, 40 degrees and 90 degrees. User can still choose another angle with the alpha parameter, then calctav will run instead of using the default transmissivity for an angle of 40 degrees (and thank you for the suggestion, it speeds up the computation, indeed!)
- OptDomain_RT includes optimal spectral domains defined in [Féret et al.(2019)](https://doi.org/10.1016/j.rse.2018.11.002)) and (Spafford et al. (2021))[https://doi.org/10.1016/j.rse.2020.112176]
