### 2024-01-04 jbferet/prospect#5 @jbferet
I decided to get rid of the function plotinv in prospect v1.5.0.  randomcoloR is now unnecessary.
Sorry again for the difficulties you had to install v8, @earth-chris. I guess you were not the only one, but the issue was not raised.
I may add support for plots in a next version, but will be more careful with the packages involved.
