### 2024-01-04 jbferet/prospect#6 @jbferet
v1.5.0: suggestion accepted
