### 2024-01-04 jbferet/prospect#7 @jbferet
v1.5.0: automated unit tests were added for the functions _PROSPECT_, _Invert_PROSPECT_ and _Invert_PROSPECT_OPT_
