### 2024-01-04 jbferet/prospect#8 @jbferet
v1.5.0: a dedicated CONTRIBUTING.md and  CODE_OF_CONDUCT.md were added.
