### 2024-01-04 jbferet/prospect#10 @jbferet
@earth-chris thank you for raising this issue. This should now be fixed in prospect 1.5.0.
The code and web documentation are now up-to-date. Please tell me if any error remains.
