### 2024-01-04 jbferet/prospect#9 @jbferet
the paper has been updated and the introduction now includes a brief history on prospect development
