### 2024-01-04 eb05c5b
minor update of the code, based on recommendations from package goodpractices. minor update of the JOSS paper
