+++ data-raw/DATASET.R
@@ -6,5 +6,6 @@ ANT <- 2*runif(10)
 EWT <- 0.04*runif(10)
 LMA <- 0.02*runif(10)
 N <- 1+2*runif(10)
-Input_PROSPECT <- data.frame('CHL'=CHL,'CAR'=CAR,'ANT'=ANT,'EWT'=EWT,'LMA'=LMA,'N'=N)
+Input_PROSPECT <- data.frame('CHL'=CHL,'CAR'=CAR,'ANT'=ANT,
+                             'EWT'=EWT,'LMA'=LMA,'N'=N)
 
