+++ R/SpecPROSPECT_FullRange-data.R
@@ -20,6 +20,8 @@
 #'  - *SAC_LMA*: specific absorption coefficient for dry matter (Leaf Mass per Area)
 #'  - *SAC_PROT*: specific absorption coefficient for proteins
 #'  - *SAC_CBC*: specific absorption coefficient for carbon-base constituents
-#'  - *calctav_90*: transmissivity of a dielectric plane surface, averaged over all directions of incidence and over all polarizations for solid angle of 90 degrees
-#'  - *calctav_40*: transmissivity of a dielectric plane surface, averaged over all directions of incidence and over all polarizations for solid angle of 40 degrees
+#'  - *calctav_90*: transmissivity of a dielectric plane surface, averaged over
+#'  all directions of incidence & all polarizations for solid angle of 90 deg
+#'  - *calctav_40*: transmissivity of a dielectric plane surface, averaged over
+#'  all directions of incidence & all polarizations for solid angle of 40 deg
 "SpecPROSPECT_FullRange"
