### 2024-01-04 jbferet/prospect#11 @jbferet
hi @earth-chris,
thank you for this review and relevant suggestions. please find below my comments / answers to your suggestions.

High level:
- yes, this R package prospect aims providing future versions of the model when developed. the paper now mentions that versions 4 and 5 prior to prospect-D and PRO are deprecated. versions in other languages are still valid, as long as they use PROSPECT-D or PROSPECT-PRO.

- the updated version of PROSPECT now includes versions D and PRO only. Brown pigments can be used with both versions. The default value is 0, and users need to set Est_Brown_Pigments = TRUE to account for brown pigments during inversion (default = FALSE).

- limitations of the software: PROSPECT focuses on leaf scale, and leaf spectroscopy analysis. Therefore, it cannot be used directly to assess vegetation properties from satellite data. However, it can be used coupled with SAIL, DART, or any canopy model, to simulate canopy reflectacne and analyze multi/hyperspectral data.

- the 1 nm spectral sampling over the full VSWIR domain is the default spectral configuration. However, users can modify the spectral constants provided in SpecPROSPECT_FullRange to adjust these spectral characteristics.

- PROSPECT inversion is usually performed on limited amount of samples from experimental databases (usually 100-1000 samples, some large leaf spectral databases may include several thousands of spectra). This is now mentioned in the paper. To my knowledge, there are no leaf optical properties databases reaching a million samples. If this case occurs, HPC should easily allow computation for such dataset.

- The updated JOSS manuscript still includes code blocks. It seemed relevant to me when writing the initial version of the paper, but I agree to remove the code examples if irrelevant. I removed the code block corresponding to the computation of the LUT, and added a function to download the ANGERS dataset in prospect v1.5.1. Thank you for noticing this!

Comments:
-     lines 24-26: "[the package] includes parameterizable inversion routines, allowing users to design and test their own inversion strategy." how can users design and test their own inversion strategy? this is not documented in the manuscript or the web docs.
--> the online documentation has been updated, and now includes an example of user defined merit function

- lines 60-65: a LUT simulation is provided but not used elsewhere in the manuscript or the web docs. It's not clear what this does, or why it is useful. An example figure, code snippet, or citation showing how LUTs are used would help.
--> the LUT simulation was removed from the manuscript

- line 73: "various inversion strategies can be found in the literature." please add citations for this.
--> added references in the JOSS manuscript

- line 82: the ANGERS download code block is extra bulky and not particularly relevant to the manuscript. i'd recommend either creating simple helper function to download this data or removing the code blocks from this section entirely.
--> Thank you for this suggestion. Good idea, indeed. A function downloading leaf databases available from my personal data repository has been added.

- lines 106-111: the conclusions state that PROSPECT can be run with bidirectional or reflectance/transmission-only data. This seems like an important point to make, and you might clarify the pros and cons of each approach. Reading the manuscript and the examples, I assumed you needed bidirectional data.
--> PROSPECT simulates directional-hemispherical reflectance and transmittance. PROSPECT inversion can run using either directional-hemispherical reflectance and transmittance, or directional-hemispherical reflectance only, or directional-hemispherical transmittance only. Some tests performed with bidirectional reflectance (collected with a contact probe) showed consistent assessment for leaf chemistry (see [Wang et al. (2021)](https://doi.org/10.1016/j.rse.2023.113614) for example). However, this is not strictly in the domain of validity of PROSPECT.
