### 2024-01-04 bc81ef8
# prospect v1.5.0 version following reviews for publication in JOSS
## addition
- added a function download_LeafDB to get leaf databases from gitlab repository

## Changes
- modified defaulty merit function as one function instead of two
- updated documentation
- updated function FitSpectralData
