+++ paper/paper.md
@@ -121,18 +121,22 @@ iterative optimization is the most widespread method to invert PROSPECT
 and assess leaf chemistry and structure from their optical properties. 
 It usually takes less than 1 second to perform PROSPECT inversion, which is acceptable
 when processing experimental datasets of leaf optical properties including 100-1000 samples.
-Iterative optimization aims at minimizing a cost function comparing 
+Iterative optimization aims at minimizing a merit function comparing 
 measured and simulated leaf optical properties. 
 This procedure is based on the function `fmincon` included in the package `pracma`.
 
-Various inversion strategies using iterative optimization are described in the literature. 
-These inversion strategies differ either by the cost function, or by the 
+Various inversion strategies using iterative optimization are described in the 
+literature [@colombo_estimation_2008, @li_retrieval_2011, @feret2019]. 
+These inversion strategies differ either by the merit function, or by the 
 selection of specific spectral domains used to retrieve one or several leaf 
 biophysical properties, or by the introduction of prior information. 
-The default cost function, `CostVal_RMSE`, corresponds to the root mean square 
+The default merit function, `Merit_PROSPECT_RMSE`, corresponds to the root mean square 
 of the mean quadratic difference between measured and simulated leaf optical 
 properties (reflectance and/or transmittance).
-Users can define their own cost function. 
+Users can define their own merit function, as long as their function uses the 
+same input and output variables as `Merit_PROSPECT_RMSE`. 
+The online documentation of the package provides an example of alternative merit function.
+
 Table \ref{table:2} provides information on the optimal spectral range used to 
 assess leaf chemical constituents from their optical properties, as identified 
 by [@feret2019; @spafford2021].
@@ -154,8 +158,6 @@ ANT: anthocyanins; BROWN: brown pigments; EWT: equivalent water thickness;
 LMA: leaf mass per area; PROT: proteins; CBC: carbon based constituents).\label{table:2}
 
 # Example 1: running PROSPECT in forward mode
-
-## Individual simulations
 
 PROSPECT simulates leaf directional-hemispherical reflectance and transmittance 
 from the leaf structure parameter and a combination of chemical constituents 
@@ -209,54 +211,17 @@ LRT_VNIR <- PROSPECT(SpecPROSPECT = VNIR$SpecPROSPECT,
                      N = 1.4, CHL = 30, CAR = 6, EWT = 0.02, LMA = 0.01)
 ```
 
-## Simulation of a Look-Up-Table
-
-Look-Up-Tables (LUTs) are widely used in order to infer leaf characteristics from PROSPECT. 
-The function `PROSPECT_LUT` computes a LUT based on a list of input parameters.
-Undefined parameters are set to their default value. Vectors of values are expected 
-to be the same length.
-The output of `PROSPECT_LUT` is a list containing a data frame including the 
-input parameters, a reflectance data frame and a transmittance data frame.
-
-```r
-# define input parametrs for PROSPECT
-Input_PROSPECT <- data.frame('CHL' = 100*runif(1000), 
-                             'CAR' = 25*runif(1000), 
-                             'ANT' = 2*runif(1000), 
-                             'EWT' = 0.04*runif(1000), 
-                             'LMA' = 0.02*runif(1000), 
-                             'N' = 1+2*runif(1000))
-# produce a LUT defined over the VSWIR domain covered by PROSPECT
-LUT <- PROSPECT_LUT(Input_PROSPECT = Input_PROSPECT)
-# produce a LUT defined over the VNIR domain
[…]
