+++ paper/paper.bib
@@ -14,6 +14,23 @@
 	pages = {542--547},
 }
 
+
+@article{colombo_estimation_2008,
+	title = {Estimation of leaf and canopy water content in poplar plantations by means of hyperspectral indices and inverse modeling},
+	volume = {112},
+	issn = {00344257},
+	url = {http://linkinghub.elsevier.com/retrieve/pii/S0034425707004257},
+	doi = {10.1016/j.rse.2007.09.005},
+	language = {en},
+	number = {4},
+	urldate = {2017-07-06},
+	journal = {Remote Sensing of Environment},
+	author = {Colombo, R and Meroni, M and Marchesi, A and Busetto, L and Rossini, M and Giardino, C and Panigada, C},
+	month = apr,
+	year = {2008},
+	pages = {1820--1834},
+}
+
 @article{feret2008,
 	title = {{PROSPECT}-4 and 5: {Advances} in the leaf optical properties model separating photosynthetic pigments},
 	volume = {112},
@@ -95,21 +112,19 @@
 	pages = {75--91},
 }
 
-@article{spafford2021,
-	title = {Spectral subdomains and prior estimation of leaf structure improves {PROSPECT} inversion on reflectance or transmittance alone},
-	volume = {252},
-	issn = {00344257},
-	url = {https://linkinghub.elsevier.com/retrieve/pii/S0034425720305496},
-	doi = {10.1016/j.rse.2020.112176},
-	language = {en},
-	urldate = {2020-11-16},
+@article{jacquemoud_prospect+_2009,
+	title = {{PROSPECT}+ {SAIL} models: {A} review of use for vegetation characterization},
+	volume = {113},
+	shorttitle = {{PROSPECT}+ {SAIL} models},
+	url = {http://www.sciencedirect.com/science/article/pii/S0034425709000765},
+	doi = {doi:10.1016/j.rse.2008.01.026},
+	urldate = {2016-01-03},
 	journal = {Remote Sensing of Environment},
-	author = {Spafford, Lynsay and le Maire, Guerric and MacDougall, Andrew and de Boissieu, Florian and Féret, Jean-Baptiste},
-	month = jan,
-	year = {2021},
-	pages = {112176},
+	author = {Jacquemoud, Stéphane and Verhoef, Wout and Baret, Frédéric and Bacour, Cédric and Zarco-Tejada, Pablo J. and Asner, Gregory P. and François, Christophe and Ustin, Susan L.},
+	year = {2009},
+	pages = {S56--S66},
+	file = {[PDF] from csic.es:C\:\\Users\\Feret\\Zotero\\storage\\76ATRJNN\\Jacquemoud et al. - 2009 - PROSPECT+ SAIL models A review of use for vegetat.pdf:application/pdf;Snapshot:C\:\\Users\\Feret\\Zotero\\storage\\UX33MX6F\\S0034425709000765.html:text/html},
 }
-
 
 @article{jay_physically-based_2016,
 	title = {A physically-based model for retrieving foliar biochemistry and leaf orientation using close-range imaging spectroscopy},
@@ -126,21 +141,35 @@
 	pages = {220--236},
 }
 
-
-@article{jacquemoud_prospect+_2009,
-	title = {{PROSPECT}+ {SAIL} models: {A} review of use for vegetation characterization},
-	volume = {113},
-	shorttitle = {{PROSPECT}+ {SAIL} models},
-	url = {http://www.sciencedirect.com/science/article/pii/S0034425709000765},
-	doi = {doi:10.1016/j.rse.2008.01.026},
+@article{li_retrieval_2011,
[…]
