+++ man/Merit_PROSPECT_RMSE.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/Lib_PROSPECT_Inversion.R
-\name{Merit_RMSE_PROSPECT}
-\alias{Merit_RMSE_PROSPECT}
+\name{Merit_PROSPECT_RMSE}
+\alias{Merit_PROSPECT_RMSE}
 \title{Merit function for PROSPECT inversion}
 \usage{
-Merit_RMSE_PROSPECT(
+Merit_PROSPECT_RMSE(
   x,
   SpecPROSPECT,
   Refl,

[…]
