+++ vignettes/prospect2.Rmd
@@ -26,33 +26,50 @@ knitr::opts_chunk$set(
 )
 ```
 
-# Running `prospect` in inverse mode: generalities
+# `prospect` inversion: generalities
 
 `PROSPECT` can be inverted using iterative optimization, by calling the function `Invert_PROSPECT`. 
 This iterative optimization is based on the function [`fmincon` included in the package `pracma`](https://rdrr.io/cran/pracma/man/fmincon.html).
 
-By default, the merit function used for the inversion minimizes the RMSE between the simulated and the measured leaf optical properties.
-However, users can define their own merit function with associated criterion to minimize by defining their own merit function and adding it as input variable, such as `MeritFunction = MyOwnMeritFunction`.
+By default, the merit function used for the inversion named [Merit_PROSPECT_RMSE](https://jbferet.gitlab.io/prospect/reference/Merit_PROSPECT_RMSE.html) 
+minimizes the RMSE between the simulated and the measured leaf optical properties.
+However, users can define their own merit function with associated criterion to minimize 
+and define it as input variable for `Invert_PROSPECT`: `MeritFunction = "MyOwnMeritFunction"`.
 
-The results of inversion will also depend on the prospect version, i.e. on the parameters included in the prospect model. 
-The following table shows the correspondence between prospect versions and prospect parameters. 
-Each parameter can be considered as parameter to estimate or a fix value in the model, depending on the prior knowledge. 
+As optimal spectral domains were identified when inverting PROSPECT with 
+`Merit_PROSPECT_RMSE`, the function `Invert_PROSPECT_OPT` does not accept other 
+merit functions. 
+However, users can easily develop their own script to identify optimal spectral 
+domains with the merit function of their choice. 
 
-| Version  | 5                                       | 5B                                    | D                                      | DB                                     | PRO                                    | PROB
-| :------: |:--------------------------------------:|:--------------------------------------:|:--------------------------------------:|:--------------------------------------:|:--------------------------------------:|:---------------------------------------:|
-| CHL      |`r emojifont::emoji('white_check_mark')`|`r emojifont::emoji('white_check_mark')`|`r emojifont::emoji('white_check_mark')`|`r emojifont::emoji('white_check_mark')`|`r emojifont::emoji('white_check_mark')`|`r emojifont::emoji('white_check_mark')`|
-| CAR      |`r emojifont::emoji('white_check_mark')`|`r emojifont::emoji('white_check_mark')`|`r emojifont::emoji('white_check_mark')`|`r emojifont::emoji('white_check_mark')`|`r emojifont::emoji('white_check_mark')`|`r emojifont::emoji('white_check_mark')`|
-| ANT      |                                        |                                        |`r emojifont::emoji('white_check_mark')`|`r emojifont::emoji('white_check_mark')`|`r emojifont::emoji('white_check_mark')`|`r emojifont::emoji('white_check_mark')`|
-| BROWN    |                                        |`r emojifont::emoji('white_check_mark')`|                                        |`r emojifont::emoji('white_check_mark')`|                                        |`r emojifont::emoji('white_check_mark')`|
-| EWT      |`r emojifont::emoji('white_check_mark')`|`r emojifont::emoji('white_check_mark')`|`r emojifont::emoji('white_check_mark')`|`r emojifont::emoji('white_check_mark')`|`r emojifont::emoji('white_check_mark')`|`r emojifont::emoji('white_check_mark')`|
[…]
