### 2024-01-06 e5d2b91
update readme, paper and get started
