+++ README.md
@@ -2,7 +2,7 @@
 
 # An R package for the simulation of leaf optical properties based on their biochemical and biophysical properties using the PROSPECT leaf model. 
 
-[![licence](https://img.shields.io/badge/Licence-GPL--3-blue.svg)](https://www.r-project.org/Licenses/GPL-3)
+[![licence](https://img.shields.io/badge/Licence-MIT-blue.svg)](https://www.r-project.org/Licenses/MIT)
 [![Build Status](https://gitlab.com/jbferet/prospect/badges/master/pipeline.svg)](https://gitlab.com/jbferet/prospect/pipelines/latest)
 
 # 1 Install
