### 2024-01-06 e420607
correct invalid reference in paper.bib
