+++ paper/paper.bib
@@ -117,13 +117,12 @@
 	volume = {113},
 	shorttitle = {{PROSPECT}+ {SAIL} models},
 	url = {http://www.sciencedirect.com/science/article/pii/S0034425709000765},
-	doi = {doi:10.1016/j.rse.2008.01.026},
+	doi = {10.1016/j.rse.2008.01.026},
 	urldate = {2016-01-03},
 	journal = {Remote Sensing of Environment},
 	author = {Jacquemoud, Stéphane and Verhoef, Wout and Baret, Frédéric and Bacour, Cédric and Zarco-Tejada, Pablo J. and Asner, Gregory P. and François, Christophe and Ustin, Susan L.},
 	year = {2009},
 	pages = {S56--S66},
-	file = {[PDF] from csic.es:C\:\\Users\\Feret\\Zotero\\storage\\76ATRJNN\\Jacquemoud et al. - 2009 - PROSPECT+ SAIL models A review of use for vegetat.pdf:application/pdf;Snapshot:C\:\\Users\\Feret\\Zotero\\storage\\UX33MX6F\\S0034425709000765.html:text/html},
 }
 
 @article{jay_physically-based_2016,
