### 2024-01-06 #6027 @jbferet
Dear @kthyng, @ashiklom , @earth-chris and @nehakn ,
After a few commits and new versions 1.5.0 and 1.5.1 of the package, all issues are now solved and missing elements or modifications suggested in your reviews have been implemented. The latest version of the package is functional. I went through the online documentation, and tried my best to improve the manuscript. Automated tests were also added to the package.

Thank you for all these excellent suggestions. Let us know if you are ok with our answers to the issues opened in the github repository, and we will close them if ok for you, or work more on it if needed!

best regards
