### 2024-01-07 jbferet/prospect#10 @earth-chris
thanks for the fix! i can execute the example code with no issues.
