### 2024-01-07 jbferet/prospect#11 @earth-chris
the new manuscript looks great! and sorry for the confusion regarding the inversion - I so often think about the inversion when applied to canopy reflectance that I misunderstood the use case here in asking my question.
