### 2024-01-11 #6027 @jbferet
@kthyng all items from the list are checked, everything should be ok on my side. The DOI for the zenodo release is in the previous message.
Thank you!
