### 2024-01-11 #6027 @jbferet
> @jbferet and what version did you want associated with this JOSS review?

@kthyng prospect version 1.6.0 is the version including all changes related to the JOSS review. Is there an action on my side in addition to specify it?
