### 2024-01-11 #6027 @kthyng
@jbferet paper comments:
* remove "[]" from around inline citations, this needs to be fixed many places, starting at line 29. See example paper for more information: https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography
* weird line break on line 50 — seems like this should be one sentence combined with the following line.
* line 65 has two sets of [] apparently
* line 67 add space
* add code formatting line 124 though installation instructions should be in the docs and in the readme, not in the paper

Please correct all these items and given the number of mistakes found at this point, please go through and do a thorough reading to check for more.
