### 2024-01-13 cc6da64
final corrections of the paper suggested by JOSS editor
