+++ paper/paper.md
@@ -24,11 +24,11 @@ bibliography: paper.bib
 
 # Summary
 
-PROSPECT simulates optical properties for a leaf based on its content
+PROSPECT simulates optical properties for a leaf, based on its content
 in absorbing constituents, and a structure parameter accounting for light scattering.
 We present `prospect`, an R package aiming at providing the most recent versions 
-of the model, along with inversion routines to assess leaf 
-chemistry from optical properties.
+of the model, along with inversion routines to assess leaf chemistry from 
+optical properties.
 
 # Statement of need
 
@@ -51,8 +51,9 @@ absorption coefficients.
 It uses a simplified representation of leaf structure through the generalised 
 plate model [@allen1970] to account for scattering. 
 PROSPECT can be coupled with canopy reflectance models to analyze Earth 
-observation data. Hence, it is a key component for 
-remote sensing applications dedicated to vegetation monitoring. 
+observation data. 
+Hence, it is a key component for remote sensing applications dedicated to 
+vegetation monitoring. 
 
 Multiple versions have been released since its first version [@jacquemoud1990].
 [@feret2008] introduced carotenoids and [@feret2017] introduced anthocyanins, 
@@ -60,8 +61,8 @@ to simulate leaf optical properties from juvenile to mature and senescent
 development stages. 
 [@feret2021] introduced PROSPECT-PRO, the latest version separating dry matter 
 constituents into proteins and carbon based constituents. 
-In parallel with the updated versions of the model, model inversion strategies 
-were introduced to improve the assessment of leaf chemical constituents 
+In parallel with updated versions of the model, model inversion strategies 
+have been introduced to improve the assessment of leaf chemical constituents 
 [@feret2019; @spafford2021].
 
 PROSPECT implementations since [@feret2008] can be found at 
@@ -90,7 +91,7 @@ Two versions of PROSPECT are implemented in `prospect`: PROSPECT-D
 [@feret2017] and PROSPECT-PRO [@feret2021]. 
 PROSPECT-4 and 5 are deprecated. 
 Table \ref{table:1} lists the input leaf chemical constituents and structure parameter
-for versions D and PRO of the function `PROSPECT`.
+for versions D and PRO available with the function `PROSPECT`.
 
 | Version 	|  `D` 	| `PRO` |
 |---------	|:----:	|:----:	|
@@ -109,7 +110,7 @@ package (CHL: chlorophylls; CAR: carotenoids; ANT: anthocyanins; BROWN: brown pi
 EWT: equivalent water thickness; LMA: leaf mass per area; PROT: proteins; 
 CBC: carbon based constituents; N: leaf structure parameter).\label{table:1}
 
-PROSPECT is a relatively simple and computationally efficient model.  
+PROSPECT is a relatively simple and computationally efficient model. 
 Hence iterative optimization can be considered to invert PROSPECT 
 and assess leaf chemistry and structure from their optical properties. 
 Iterative optimization aims at minimizing a merit function comparing 
@@ -130,7 +131,7 @@ Alternative merit functions using the same input and output variables as
 The online documentation provides an example of alternative merit function.
 
 Table \ref{table:2} details the optimal spectral domains to assess leaf chemical 
-constituents from their optical properties ([@feret2019; @spafford2021]).
+constituents from their optical properties [@feret2019; @spafford2021].
 
 | Constituent|    Optimal spectral domain     	|   Versions   |
 |----------- |---------------------------------	|:-----------: |
[…]
