### 2024-01-16 e28cca4
# prospect v1.6.1 ## addition - added input dataframe InputPROSPECT to function PROSPECT
