+++ man/define_Input_PROSPECT.Rd
@@ -1,0 +1,51 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/Lib_PROSPECT.R
+\name{define_Input_PROSPECT}
+\alias{define_Input_PROSPECT}
+\title{This function produces a data frame from all prospect input variables if not
+defined already}
+\usage{
+define_Input_PROSPECT(
+  Input_PROSPECT,
+  CHL = NULL,
+  CAR = NULL,
+  ANT = NULL,
+  BROWN = NULL,
+  EWT = NULL,
+  LMA = NULL,
+  PROT = NULL,
+  CBC = NULL,
+  N = NULL,
+  alpha = NULL
+)
+}
+\arguments{
+\item{Input_PROSPECT}{numeric. content corresponding to LMA}
+
+\item{CHL}{numeric. Chlorophyll content (microg.cm-2)}
+
+\item{CAR}{numeric. Carotenoid content (microg.cm-2)}
+
+\item{ANT}{numeric. Anthocyanin content (microg.cm-2)}
+
+\item{BROWN}{numeric. Brown pigment content (Arbitrary units)}
+
+\item{EWT}{numeric. Equivalent Water Thickness (g.cm-2)}
+
+\item{LMA}{numeric. Leaf Mass per Area (g.cm-2)}
+
+\item{PROT}{numeric. protein content  (g.cm-2)}
+
+\item{CBC}{numeric. NonProt Carbon-based constituent content (g.cm-2)}
+
+\item{N}{numeric. Leaf structure parameter}
+
+\item{alpha}{numeric. Solid angle for incident light at surface of leaf}
+}
+\value{
+list. updated LMA, PROT and CBC
+}
+\description{
+This function produces a data frame from all prospect input variables if not
+defined already
+}
