### 2024-01-19 #6027 @kthyng
Here is a PR to deal with the inline references: https://github.com/jbferet/prospect/pull/12
