### 2024-02-02 d948127
updated Florian's contact in DESCRIPTION file
