### 2024-02-05 #6027 @kthyng
@jbferet I am still seeing those references as parenthetical instead of inline. Can you tell what is wrong? I see you merged the PR but I don't see the updates in the history of the repository. Did the default branch change?
