+++ man/Merit_PROSPECT_dRMSE.Rd
@@ -1,0 +1,36 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/Lib_PROSPECT_Inversion.R
+\name{Merit_PROSPECT_dRMSE}
+\alias{Merit_PROSPECT_dRMSE}
+\title{Merit function for PROSPECT inversion}
+\usage{
+Merit_PROSPECT_dRMSE(
+  x,
+  SpecPROSPECT,
+  Refl,
+  Tran,
+  Input_PROSPECT,
+  Parms2Estimate
+)
+}
+\arguments{
+\item{x}{numeric. Vector of input variables to estimate}
+
+\item{SpecPROSPECT}{list. Includes optical constants refractive index,
+specific absorption coefficients and corresponding spectral bands}
+
+\item{Refl}{numeric. measured reflectance data}
+
+\item{Tran}{numeric. measured Transmittance data}
+
+\item{Input_PROSPECT}{dataframe. set of PROSPECT input variables}
+
+\item{Parms2Estimate}{numeric. location of variables from Input_PROSPECT
+to be estimated through inversion}
+}
+\value{
+fc estimates of the parameters
+}
+\description{
+Merit function for PROSPECT inversion
+}
